[fix] card_150 モノクロ機能実装
* Filterファイルわけ
